Add API call to move gateways onto the fabric
If we don't handle all gateways we might want to move gateways in
certain AZs (or even one by one in an AZ) onto the fabric. This can now
be done via a network API call! We emit an event for
"cc-fabric-network-gateway" with BEFORE_UPDATE, so other drivers
currently handling the l3 gateway for that network can tear it down
before we set it up.