Ensure subnet-network AZ and subnetpool AZ match
When a new subnet is created in an external network we now check that
the subnet's network's AZ hint matches the AZ of the subnetpool (given
to the subnetpool by a tag). This check is only executed if the
subnetpool has an address scope, because a) we only need to enforce this
for networks that are also "present in a VRF", which is managed via
address scopes (else the subnet is really more of an internal flavor...)
and b) because it was a bit more convenient to implement for me (which
means that we can definitely change this implementation if we need to).

The config option subnet_subnetpool_az_check_enabled is added so we can
turn this feature off if it ever breaks something in prod and we quickly
need to disable it without doing a drive code change. I don't really
expect us needing it, but you never know.