L3 SVI/BGP on BGWs/Transits + binding rework
We now setup l3 on BGWs and Tranists (i.e. all interconnects). This
means that we have to do the same SVI (vlan_ifaces) and BGP VRF setup
that we need to do everywhere else when an l3 port is bound.

This whole thing put me in front of the question "why do we have 13
places that all call add_binding_host_to_config() and do something only
slightly different"? As all of these places need similar l3 config, I
decided to rework most of the code that calls
add_binding_host_to_config() to now mostly call
add_segments_to_config(), which then pulls the necessary l3 config and
properly fills the SwitchConfigUpdateList with this. This was a medium
sized refactoring, which will hopefully make the whole code much more
maintainable in the future.