[netbox-model] Only log and ignore if name is inconsistent
Whenever we encounter a device that does not fit our naming convention,
we raise an exception and terminate. That's tedious as we need to poke
people to correct all their switch names when we run this script.
Instead, we should just chose to ignore every device that does not
comply with the naming scheme and move on.