Reset oslo.config overrides for all tests
Config set in oslo.config with set_override() is not necessarily removed
when not explicitly overridden in a test and can cause other tests in
the same worker to fail due to leftover overrides. We now use a fixture
to reset overrides done by the tests, as it is done in the tests of some
Neutron ml2 drivers and also in general in Nova, where we got the
inspiration from.

https://github.com/sapcc/nova/blob/6d0d19c069d1ea70fbed225ddb9f68f78d271a03/nova/test.py#L210