Move GNMI client out of EOS agent, make it common
EOS was the only agent where we planned on using GNMI, but now it looks
like we're going to use GNMI with NXOS as well. Therefore the GNMI
client abstraction we have moves into common/ and all agents that want
to use it can use it. The agents will keep their vendor specific stuff
(like paths or regexes) in their own code base though.