Tests: Set MTU for networks used in tests.
Since this change[0] the MTU set by _make_network() is set to the
maximum, not to nl_const.DEFAULT_NETWORK_MTU anymore. As some
TypeDrivers reserve some bytes for their tunneling they don't offer the
full maximum MTU. This is only a problem for tests not instantiating a
ml2 driver, meaning our MechanismDriver tests are working fine, the
other tests creating networks are not. In these other tests we also
don't have the option to set the MTU via API, as the API does not
recognize the attribute, most likely due to a missing default plugin.

The easiest way for me to get the tests running again is to directly
update the MTU in the DB after network creation. Might change in the
future, when we find a better way to do this.

[0] Ica21e891cd2559942abb0ab2b12132e7f6cdd835 or
https://github.com/openstack/neutron/commit/fec0286e26b5c0b02f1ebdc8d3d481f73fbe4aec