Fix TypeError on missing network in update_port_postcommit()
When a port is updated we try to clean up after it. As there is no
explicit delete message, but only a update_port_postcommit() call,
we handle deletion of the old host. For this, the delete call also needs
the network, but in certain scenarios, there is no original network -
most likely in cases where the port only gets its binding host updated,
but stays in the same network. Therefore in these cases we now use the
network_id given in the original bindings level. As segments are not
always fully specified (or better: this is what I have seen in some
places in the code, but am unsure if the segments are always fully
specified or not within a NetworkContext (fully specified in this case
means "missing the network_id entry")), I chose the "safe way", using
the original network first, then look at the segment dict and if that
fails we log a warning and return, i.e. don't try to clean up after the
segment.

This fixes a TypeError, where when there was no original network, we'd
try to lookup ['id'] to the original network, which was None.