More TypeError fixing in update_port_postcommit()
In some cases update_port_postcommit() gets two old segments, but the
second segment can be None, apparently. This, again, leads us to a
TypeError, as we try to find out if 'network_id' is in None.

Well, now we check if the segment is none beforehand and even make the
core a bit shorter by giving the original bindings a name.