EOS: Convert VLAN UUID names to 32 chars
VLANs on EOS can have a maximum length of 32 characters. As we're using
the network id as VLAN name in many places we now convert UUIDs to 32
chars by removing the "-" in it. To be able to diff the config later on
we also transform this back to a UUID when we read the config by adding
"-" in the appropriate places when the VLAN name is a 32 char hex
string.

As OpenStack ids are lower-case and we are mainly interested in
converting names back that we put there ourselves we at the moment do
the conversion only for lower-case UUIDs/characters. Let's see if that
gets problematic in the future.