Implement trunk extension
Implement the trunk service extension. A trunk can be created on any
port that is in a direct binding hostgroup. Only one trunk per hostgroup
is possible. If the host is part of a metagroup, only VLANs not used by
the metagroup VLAN pool can be used as target, to not interfere with VMs
put on the host(s).

The trunk plugin modifies the subports by setting binding host, vnic
type and binding profile. The binding profile contains details
about the specific VLAN translation, but this is only informational, as
the "real" info will be fetched from the respective trunk/subport tables
in the DB.