Only consider own subports for trunking
When querying segment info from the DB we also query out the trunk id
for subports. Trunking for the OpenStack trunking extension always needs
to happen at the last driver, so in this case we must ignore subports
that belong to other drivers. We will only handle trunking for our own
ports, this being ports that are bound by us as the last driver. We
identify these ports by their vif_type, cc-fabric.