Allow hostgroups to bind multiple trunks w/ config
If a hostgroup has allow_multiple_trunk_ports set it can now have
multiple trunks on it. This allows us to have multiple domains/projects
set VLAN translations for the same device without the networks needing
to be shared in to the same project. As we don't know in this case which
port should dictate the native VLAN we don't set the native VLAN at all
for these groups.