Make port speed configurable (in EOS as well)
As the driver is fully responsible for its own switchports, we use a
replace config call via GNMI to configure those ports. This means that
there is - at the moment - no option for manual configuration of these
interfaces. This means that we now also need to be able to configure
interface port speeds.

Config-wise this means that we just have a string, which might be
vendor-specific, but at least understandable by the platform's agent. In
the eos case we translate speeds like 10/100g to their respective API
values (SPEED_10GB, SPEED_100GB).