EOS GNMI compatibility for EOS >= 4.32
Somewhere between EOS 4.28.9M and 4.32.3M the GNMI structure for
vlan-vni mappings and for aggregates changed.

For vlan-vni mappings the path changed, so we now need to check on which
version we are and use the corresponding path for querying and
configuring them. It's a bit weird now as each vlan element has a config
and a state entry, but it looks like EOS just wants to introduce more
state reporting into their GNMI API.

For the aggregates, we now need to filter both for protocol name and
identifier name. In case of BGP they're both BGP. Querying works even
without specifying the identifier, but configuration does not work
anymore. Adding [identifier=BGP] to all of them solves this problem.

With these two changes we now need to implement a version switch. For
now the version is fetched on first use and gets reset on each connect,
so we make sure that we pick up new versions if we loose the GNMI
connection (as this could very well be caused by a switch firmware
update!). The version check is a rather simple tuple conversion /
comparison and is currently based on the assumption that the first two
parts of the version number will always be plain numeric (like 4.32.3M).